修改了关卡配置信息，增加了挂机关卡配置 成就数据库表修改了数据类型 工厂模块做了优化更新 玩家创建角色初始赠送做了部分修改 方法访问的延迟做了部分修改 关卡模块做了部分新方法适配，暂为完成

+++ utility/world_creator.py
@@ -27,9 +27,9 @@ ACHIEVEMENT = \
 """
 CREATE TABLE `achievement` (
 	 `uid` VARCHAR(128) COLLATE utf8mb4_unicode_ci NOT NULL COMMENT '用户唯一id',
-	 `aid` SMALLINT UNSIGNED NOT NULL COMMENT '成就唯一id',
-	 `value` SMALLINT UNSIGNED DEFAULT 0 COMMENT '成就完成的次数',
-	 `reward` SMALLINT UNSIGNED DEFAULT 0 COMMENT '完成成就领奖的次数',
+	 `aid` INT UNSIGNED NOT NULL COMMENT '成就唯一id',
+	 `value` INT UNSIGNED DEFAULT 0 COMMENT '成就完成的次数',
+	 `reward` INT UNSIGNED DEFAULT 0 COMMENT '完成成就领奖的次数',
 	 PRIMARY KEY (`uid`,`aid`),
 	 CONSTRAINT `achievement_player_1` FOREIGN KEY (`uid`) REFERENCES `player` (`uid`) ON DELETE CASCADE ON UPDATE CASCADE
 	) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
